Make sure that the purging of deleted and trashed objects only occurs when the flow operations are called from the cronjob, not when publishing.
git-svn-id: http://svn.ez.no/svn/extensions/ezflow/trunk@3775 49f50cf4-cce6-0310-bc1e-a21d999f3d8f

+++ packages/ezflow_extension/ezextension/ezflow/classes/ezflowoperations.php
@@ -129,7 +129,7 @@ class eZFlowOperations
      * 
      * @static
      */
-    public static function update()
+    public static function update( $cronjobMode = false )
     {
         include_once( 'kernel/classes/ezcontentcache.php' );
 
@@ -151,33 +151,37 @@ class eZFlowOperations
             $db->commit();
         }
 
-        // Find items that have been moved to trash or deleted
-        $itemArray = array();
-        $offset = 0;
-        $limit = 50;
-        do
+        // This code will run when called from the cronjob, and not when publishing.
+        if ( $cronjobMode )
         {
-            $items = $db->arrayQuery( 'SELECT object_id FROM ezm_pool', array( 'offset' => $offset, 'limit' => $limit ) );
-            if ( count( $items ) == 0 )
-                break;
-
-            foreach( $items as $item )
+            // Find items that have been moved to trash or deleted
+            $itemArray = array();
+            $offset = 0;
+            $limit = 50;
+            do
             {
-                // Find items that have been deleted or moved to trash
-                $rows = $db->arrayQuery( 'SELECT id, status FROM ezcontentobject WHERE id = ' . $item['object_id'] );
-                if ( count( $rows ) == 0 or // deleted
-                     ( count( $rows ) == 1 and $rows[0]['status'] == 2 ) ) // trashed
-                    $itemArray[] = $item['object_id'];
-            }
+                $items = $db->arrayQuery( 'SELECT object_id FROM ezm_pool', array( 'offset' => $offset, 'limit' => $limit ) );
+                if ( count( $items ) == 0 )
+                    break;
 
-            $offset += $limit;
-        } while ( true );
-        // Remove them all from the flow
-        if ( count( $itemArray ) > 0 )
-        {
-            $db->begin();
-            $db->query( 'DELETE FROM ezm_pool WHERE ' . $db->generateSQLINStatement( $itemArray, 'object_id' ) );
-            $db->commit();
+                foreach( $items as $item )
+                {
+                    $rows = $db->arrayQuery( 'SELECT id, status FROM ezcontentobject WHERE id = ' . $item['object_id'] );
+                    if ( count( $rows ) == 0 or // deleted
+                         ( count( $rows ) == 1 and $rows[0]['status'] == 2 ) ) // trashed
+                        $itemArray[] = $item['object_id'];
+                }
+
+                $offset += $limit;
+            } while ( true );
+
+            // Remove them all from the flow
+            if ( count( $itemArray ) > 0 )
+            {
+                $db->begin();
+                $db->query( 'DELETE FROM ezm_pool WHERE ' . $db->generateSQLINStatement( $itemArray, 'object_id' ) );
+                $db->commit();
+            }
         }
 
         // Update pool and pages for all nodes
